NF: unsuspendCards in abstract
